[FE] Feat: 앨범 상세 페이지 서버 데이터와 연결
- getServerSideProps로 api 요청 후 데이터 받아오도록 수정
- id에 해당하는 응답 데이터 없을 시 not found true 설정
